pyproject.toml
[project]
name = "django-is-awsome"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref (==3.11.0)",
    "django (==5.2.8)",
    "gunicorn (==23.0.0)",
    "packaging (==25.0)",
    "sqlparse (==0.5.3)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-yasg (>=1.21.11,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
